[#21][refactor]controller 패키지를 web으로 변경 후 mockapi 패키지에 controller 이동
