+++ src/main/java/dev/idion/bladitodo/web/mockapi/MockAPIController.java
@@ -1,10 +1,10 @@
-package dev.idion.bladitodo.controller;
+package dev.idion.bladitodo.web.mockapi;
 
-import dev.idion.bladitodo.controller.dto.BoardDTO;
-import dev.idion.bladitodo.controller.dto.BoardWithListIdResponse;
-import dev.idion.bladitodo.controller.dto.CardDTO;
-import dev.idion.bladitodo.controller.dto.ListDTO;
-import dev.idion.bladitodo.controller.dto.UserDTO;
+import dev.idion.bladitodo.web.dto.BoardDTO;
+import dev.idion.bladitodo.web.dto.BoardWithListIdResponse;
+import dev.idion.bladitodo.web.dto.CardDTO;
+import dev.idion.bladitodo.web.dto.ListDTO;
+import dev.idion.bladitodo.web.dto.UserDTO;
 import java.util.ArrayList;
 import java.util.Arrays;
 import org.springframework.web.bind.annotation.GetMapping;
